Allow drag-drop of files within QFrames
The biggest takeaway right now is that promoting an item (QtWidgets) to another class allows me to create my own classes that subclass built-in ones.

In this context, that is having a custom QFrame class that allows drag and drop of files, but not having it just for the entire window.