Attempt: main.py file (non-functional)
Result: does not work. Cannot be run normally either.

The main issue appears to be the custom class QFrameDragDrop in qframedragdrop.py. Because this is a "package," it needs to be imported as `from TeamsAtlasBridge.qframedragdrop import QFrameDragDrop`.

But pyuic5 does not do this. And the options `--import-from=TeamsAtlasBridge` and `--from-imports` do not appear to have any effect